requirements.txt
# =====================================================
# Video Brain — Dependencies
# =====================================================

# --- Web Framework ---
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
aiofiles==24.1.0

# --- AI / ML Models ---
faster-whisper==1.1.0
fastembed==0.4.0
transformers==4.46.0
torch==2.4.0
torchvision==0.19.0
Pillow==10.4.0

# --- Vector Database ---
qdrant-client==1.12.0

# --- Utilities ---
python-dotenv==1.0.1
pydantic==2.9.0
pydantic-settings==2.5.0
